removed message struct

+++ src/GameSession/GameSession.cpp
@@ -1,6 +1,6 @@
 #include "include/GameSession.h"
 
-GameSession::GameSession(int ownerID) : 
+GameSession::GameSession(uintptr_t ownerID) : 
     //_invitationCode (Invitation()), 
     _ownerID (ownerID) {}
